# stage2: Pop error trace frames for handled errors (#1923)