… src/Zir.zig
… src/arch/aarch64/CodeGen.zig
… src/arch/arm/CodeGen.zig
… src/arch/riscv64/CodeGen.zig
… src/arch/sparc64/CodeGen.zig
… src/arch/wasm/CodeGen.zig
… src/arch/x86_64/CodeGen.zig
… src/codegen/c.zig
… src/codegen/llvm.zig
… src/print_air.zig
… src/print_zir.zig
… src/value.zig
… test/behavior/bugs/12891.zig
… test/behavior/error.zig
… test/behavior/eval.zig
… test/stack_traces.zig