This implements the (unapproved) error handling behavior described in https://github.com/ziglang/zig/issues/1923#issuecomment-1218495574.

The result is a basic ownership-style algorithm for errors. Barring any bugs (🤞), it should guarantee that error return frames follow a strict stack discipline, which we need to correctly accumulate them in a single stack-per-thread.

Error return traces "live" when:
  - stored in a `const` variable
  - handled in a `catch { ... }` or `else |err| { ... }` block
  - propagated to caller via `return` or `try`
  - passed to a function as an argument

Error return traces are killed when:
   - stored in a mutable `var`
   - leaving a `catch` or `else |err|` block with a non-error value
   - the `const` variable associated with an error trace falls out of scope
   - a function receives an error as an argument and returns a non-error
   
Here's an example that tries to show most of this together:
 ```zig
fn foo() !void {
    eat_dinner() catch {
        // Begin error handling here
        find_recipe() catch {};        // find_recipe() is removed from trace upon leaving this catch block
        if (cond)
            return error.HouseOnFire;  //      error result: Includes eat_dinner() in the error return trace, but not find_recipe()
        try cook_recipe();             //      error result: If this fails, the error return trace includes eat_dinner(), followed by cook_recipe()
    }
    // We made it out of another catch block, eat_dinner() is removed from the trace 
    return error.MyInlawsAreComing;    //      error result: Error trace starts fresh here
}

fn bar() !void {
    // In addition to returning from within the block, you can pass
    // the result to an external return, try, catch, or if:
    return try_me() catch |err| switch(err) {
        error.IgnorableError => null,   // non-error result: Pops try_me() from the error return trace
        error.ImportantError => b: {
            break :b error.FatalError;  //     error result: Includes try_me() in the error return trace
        },
        else => err,                    //     error result: Includes try_me() in the error return trace
    };
}
test {
    try expectError(error.FatalError, bar());         // If this fails, `bar()` appears in the error return trace
    try expectError(error.MyInlawsAreComing, foo());  // If this fails, `foo()` appears in the error return trace (`bar` was already popped)
}
```

The main caveat is that traces are lost when storing errors to a mutable `var`. For example:
```zig
fn foo() !void {
    var x = try_me();   // If this were `const` or `return try_me();`, the trace wouldn't be killed here
     // ... fancy logic here
    return x;           // Error trace starts fresh here - try_me() is not included.
}
```

**Note:** There is room to relax this last condition for `var`, but only a little bit. We could do the same thing we do for `const` and keep the error trace alive for the remainder of the block where the _assignment_ happens. Any wider scope will violate the strict stack discipline for error return traces, so it won't work. In the end, I decided the most consistent behavior for the user is just to kill error return traces when assigning to `var`.
 
Supersedes #12825. Closes #1923.
(Many thanks @Vexu for that PR btw -- it was very helpful to tighten up some of my handling in AstGen)