@topolarity 2022-09-14
Uh oh, I found a problematic example:

```zig
const expectError = @import("std").testing.expectError;

fn alwaysErrors() !void { return error.BUG_ThisErrorShouldNotAppearInAnyTrace; }
fn foo() !void { return error.Foo; }

test "test expected error" {
    try expectError(error.BUG_ThisErrorShouldNotAppearInAnyTrace, alwaysErrors());
    try expectError(error.BUG_ThisErrorShouldNotAppearInAnyTrace, alwaysErrors());

    try expectError(error.Bar, foo());
}
```

Without this PR, the error trace is:
```console
Test [1/1] test.test expected error... expected error.Bar, found error.Foo
Test [1/1] test.test expected error... FAIL (TestExpectedError)
./test_err6.zig:3:27: 0x20f3b8 in alwaysErrors (test)
fn alwaysErrors() !void { return error.BUG_ThisErrorShouldNotAppearInAnyTrace; }
                          ^
./test_err6.zig:3:27: 0x20f3b8 in alwaysErrors (test)
fn alwaysErrors() !void { return error.BUG_ThisErrorShouldNotAppearInAnyTrace; }
                          ^
./test_err6.zig:4:18: 0x20f6c8 in foo (test)
fn foo() !void { return error.Foo; }
                 ^
/home/topolarity/repos/zig/lib/std/testing.zig:41:13: 0x20f78f in expectError__anon_1116 (test)
            return error.TestExpectedError;
            ^
./test_err6.zig:12:5: 0x20f9b2 in test.test expected error (test)
    try expectError(error.Bar, foo());
    ^
0 passed; 0 skipped; 1 failed.
```

With it, the trace is:
```console
Test [1/1] test.test expected error... expected error.Bar, found error.Foo
Test [1/1] test.test expected error... FAIL (TestExpectedError)
/home/topolarity/repos/zig/stage4/lib/zig/std/testing.zig:41:13: 0x20f756 in expectError__anon_1116 (test)
            return error.TestExpectedError;
            ^
./test_err6.zig:12:5: 0x20f9ac in test.test expected error (test)
    try expectError(error.Bar, foo());
    ^
0 passed; 0 skipped; 1 failed.
```

Before we had too many entries, now we have too few.