@topolarity 2022-09-23
With the latest push, examples in the style of https://github.com/ziglang/zig/issues/11593 work as expected:

```zig
const expectError = @import("std").testing.expectError;

fn alwaysErrors() !void { return error.BUG_ThisErrorShouldNotAppearInAnyTrace; }
fn foo() !void { return error.Foo; }

test "test expected error" {
    try expectError(error.BUG_ThisErrorShouldNotAppearInAnyTrace, alwaysErrors());
    try expectError(error.BUG_ThisErrorShouldNotAppearInAnyTrace, alwaysErrors());
    try expectError(error.Bar, foo());
}
```

Output with this PR is:
```console
Test [1/1] test.test expected error... expected error.Bar, found error.Foo
Test [1/1] test.test expected error... FAIL (TestExpectedError)
./testme.zig:4:18: 0x211d08 in foo (test)
fn foo() !void { return error.Foo; }
                 ^
/home/topolarity/repos/zig/build/stage3/lib/zig/std/testing.zig:37:13: 0x211dcf in expectError__anon_1111 (test)
            return error.TestExpectedError;
            ^
./testme.zig:9:5: 0x212052 in test.test expected error (test)
    try expectError(error.Bar, foo());
    ^
0 passed; 0 skipped; 1 failed.
```

Before we had entries from each expectError call. Now, the entries are cleaned up if `expectError` passes.