@topolarity 2022-09-26
Was able to lift the `const` restriction. This new behavior no longer requires modifying the test runner 🚀