@topolarity 2022-10-16
Alright, looks like the rebase went well

The rebase did expose an existing bug https://github.com/ziglang/zig/issues/13175 which together with this change causes 1 behavior test regression. I've skipped that test for now