@topolarity 2022-10-20
Very good questions.

The same things gave me concern - Let's see if I can get a read on the situation for you.

> How might this integrate with async functions? In stage1 we combine error return traces when doing await.

I think the strategy here can stay largely the same. Merging the traces means that it's as if the error were generated "at" the await. For the save/restore logic here, if we treat an errorable `await` exactly the way that we treat an errorable function call, then popping should continue to work correctly.

> Can you share how this affects the compilation speed of the self-hosted compiler?

This is the most significant difference I'd expect. This does add a significant number of new ZIR instructions:

```
; src/codegen/c.zig (master)
# Source bytes:       158.0126953125KiB
# Tokens:             34052 (166.29296875KiB)
# AST Nodes:          17641 (224.0751953125KiB)
# Total ZIR bytes:    554.3564453125KiB
# Instructions:       34155 (300.1904296875KiB)
# String Table Bytes: 18.626953125KiB
# Extra Data Items:   60282 (235.4765625KiB)

; src/codegen/c.zig (this PR)       
# Source bytes:       158.0126953125KiB
# Tokens:             34052 (166.29296875KiB)
# AST Nodes:          17641 (224.0751953125KiB)
# Total ZIR bytes:    585.8505859375KiB
# Instructions:       36505 (320.8447265625KiB)
# String Table Bytes: 18.626953125KiB
# Extra Data Items:   63057 (246.31640625KiB)
```

That's a ~6.9% increase in instructions, and a 5.7% increase in ZIR bytes (sorry to eat up your `.try` savings :grimacing:)

Here are the build timings:

```
# master (stage3) ReleaseSafe building Debug
54.64s user 2.51s system 101% cpu 56.450 total
54.95s user 2.28s system 101% cpu 56.521 total
55.01s user 2.26s system 101% cpu 56.559 total

# this PR (stage3) ReleaseSafe building Debug
55.86s user 2.44s system 101% cpu 57.544 total
56.30s user 2.31s system 101% cpu 57.894 total
55.33s user 2.36s system 101% cpu 56.963 total


# master (stage3) Debug building Debug
102.98s user 2.70s system 103% cpu 1:42.59 total
103.23s user 3.25s system 102% cpu 1:43.62 total
102.81s user 2.88s system 102% cpu 1:42.73 total

# this PR (stage3) Debug building Debug
103.77s user 2.92s system 102% cpu 1:43.73 total
105.92s user 2.95s system 102% cpu 1:45.86 total
104.37s user 2.74s system 102% cpu 1:43.99 total
```

Each run was done from a completely fresh cache, so that this includes AstGen and `build.zig` compilation time.

> Can you provide any benchmark data about how runtime performance is affected?

The self-compilation timings above should include the runtime performance hit, in addition to the delta from code changes in this PR. I did try to make some more adversarial benchmarks that would do a lot of save/restore, but my attempts didn't find a delta that rose above the measurement noise yet.

In general there are two places with extra runtime code:
  1. Error handling: Trace index save/restore at every block entry/exit for `catch`, `else |err|`, and any blocks containing `const x = foo();` where foo() is errorable
  2. Discarded error traces: Trace index save/restore wrapping any errorable function call whose trace is "killed" (generally, those that do not go to `try`/`catch`/`return`/`if-else-err`/`const x = ...`)

`try foo()` and `return foo()` require no extra interactions with the trace index (except when a `return` exits an error-handling block as in (1)). Overall, compared to the overhead of error return tracing in general I think this is probably not a large delta, but let me know if you'd like me to test some existing projects or continue hunting for adversarial cases.