@andrewrk 2022-10-21
Thanks for taking these measurements. I'm a little confused by "ReleaseSafe" here - did you hack up the build.zig script? Because it only exposes the `-Drelease` flag which will select `ReleaseFast`.